pyproject.toml
[project]
name = "mcp-trove-crunchtools"
version = "0.1.0"
description = "Self-hosted local file indexing MCP server with semantic search"
requires-python = ">=3.10"
license = "AGPL-3.0-or-later"
authors = [
    { name = "crunchtools.com" }
]
readme = "README.md"
keywords = ["mcp", "rag", "semantic-search", "embeddings", "sqlite"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: GNU Affero General Public License v3 or later (AGPLv3+)",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "fastmcp>=2.0",
    "pydantic>=2.0",
    "fastembed>=0.4",
    "sqlite-vec>=0.1",
    "pymupdf4llm>=0.0.17",
    "python-docx>=1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "ruff>=0.8",
    "mypy>=1.13",
]

[project.scripts]
mcp-trove-crunchtools = "mcp_trove_crunchtools:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/mcp_trove_crunchtools"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "S", "B", "C4", "C901", "DTZ", "T10", "ISC", "PIE", "PT", "RET", "SIM", "TID", "TCH", "ARG", "PLC", "PLE", "PLR", "PLW", "TRY", "RUF"]
ignore = [
    "S101",     # assert allowed in tests
    "TRY003",   # long exception messages ok
    "PLR0913",  # many arguments ok for tools
    "PLR2004",  # magic numbers ok (HTTP status codes, limits)
    "PLW0603",  # global statement ok for singletons
    "PIE790",   # pass in empty classes ok
    "PLC0415",  # imports in tests ok
    "S105",     # hardcoded passwords in tests ok
    "TID252",   # relative imports in package ok
    "RUF022",   # __all__ sorted by category, not alphabetically
]

[tool.ruff.lint.pylint]
max-branches = 10
max-statements = 50

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["S105", "PLC0415", "PLR2004", "ARG002"]

[tool.mypy]
python_version = "3.10"
strict = true
warn_return_any = true
warn_unused_configs = true

[[tool.mypy.overrides]]
module = ["fastembed", "pymupdf4llm", "docx", "sqlite_vec"]
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
